Revert "Feature/wk 227 handle form posts"
